fix: remove --load flag from preliminary Docker build
The --load flag was creating an unknown/unknown manifest entry in the
multi-platform image. Since we only need to test the build process
and don't actually need to load the image locally, removing --load
ensures we only get the intended linux/amd64 and linux/arm64 platforms
in the final published image.