Fix device_id not set error by setting client.device_id before restore_login
CRITICAL FIX: Set matrix_client.device_id before calling restore_login()

The matrix-nio library requires the device_id to be set on the client object
before calling restore_login(), not just passed as a parameter.

This fixes the 'Device id is not set' error that occurs when loading the E2EE store.